Update Futures WebSocket streams
Add current Futures WebSocket subscription channels and payload models from the Gate Futures WebSocket documentation.

Fix order book update validation, extend stats intervals, map stream enum responses, and use decimal sizes for updated Futures stream payloads.

+++ src/Gate.IO.Api/Futures/Responses/GateFuturesUserDeleverage.cs
@@ -21,13 +21,13 @@ public record GateFuturesUserDeleverage
     /// Gets or sets the Position Size.
     /// </summary>
     [JsonProperty("position_size")]
-    public long PositionSize { get; set; }
+    public decimal PositionSize { get; set; }
 
     /// <summary>
     /// Gets or sets the Trade Size.
     /// </summary>
     [JsonProperty("trade_size")]
-    public long TradeSize { get; set; }
+    public decimal TradeSize { get; set; }
 
     /// <summary>
     /// Gets or sets the Time.
